requirements.txt
# RSJ-FFMPEG Power Toolkit Requirements
# Core dependencies for ultimate multimedia automation

# Core
python>=3.8

# CLI & Interface
click>=8.0.0
rich>=10.0.0
colorama>=0.4.4

# API Server
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Bot Integration
python-telegram-bot>=13.7
discord.py>=1.7.3

# v2.1 New Dependencies
# Web Dashboard
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Performance & Monitoring
psutil>=5.9.0

# Community & Analytics
python-dateutil>=2.8.0

# Optional AI Dependencies
# opencv-python>=4.5.0  # For face detection
# mediapipe>=0.8.0      # For advanced face tracking
# torch>=1.9.0          # For GPU acceleration
# torchvision>=0.10.0   # For AI models
aiohttp>=3.7.4

# Video/Audio Processing
opencv-python>=4.5.3
Pillow>=8.3.0
numpy>=1.21.0

# AI Enhancement (Optional - for advanced features)
# torch>=1.9.0
# torchvision>=0.10.0
# realesrgan>=0.2.0
# demucs>=3.0.0

# Utilities
requests>=2.26.0
python-dotenv>=0.19.0
watchdog>=2.1.5
schedule>=1.1.0

# Data & Config
PyYAML>=5.4.1
toml>=0.10.2

# Progress & Logging
tqdm>=4.62.0
loguru>=0.5.3

# ============================================
# Version 2.2 - New Dependencies
# ============================================

# AI & NLP
openai>=1.0.0
langchain>=0.1.0
anthropic>=0.8.0

# Cloud SDKs
boto3>=1.28.0
google-cloud-storage>=2.10.0
google-cloud-video-transcoder>=1.0.0
azure-storage-blob>=12.19.0

# Analytics & Data Processing
pandas>=2.0.0
plotly>=5.17.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Audio Analysis
librosa>=0.10.0
pydub>=0.25.0

# System Monitoring
GPUtil>=1.4.0